+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Show admin contact on "About" page.
+- Display a message when gallery is empty.
 
 ### Changed